Fix warning in debug printf.

+++ cups/dest.c
@@ -3685,7 +3685,7 @@ cups_enum_dests(
     }
 
 #  ifdef HAVE_AVAHI
-    DEBUG_printf(("1cups_enum_dests: remaining=%d, browsers=%d, completed=%u, count=%u, devices count=%u", remaining, data.browsers, (unsigned)completed, (unsigned)count, (unsigned)cupsArrayGetCount(data.devices)));
+    DEBUG_printf(("1cups_enum_dests: remaining=%d, browsers=%u, completed=%u, count=%u, devices count=%u", remaining, (unsigned)data.browsers, (unsigned)completed, (unsigned)count, (unsigned)cupsArrayGetCount(data.devices)));
 
     if (data.browsers == 0 && completed == cupsArrayGetCount(data.devices))
       break;
